Move path tests to separate module
